+++ core/src/main/java/br/ufmg/dcc/latin/querying/ResultSet.java
@@ -3,13 +3,13 @@ package br.ufmg.dcc.latin.querying;
 public class ResultSet{
 
 	public int[] docids;
-	public float[] scores;
+	public double[] scores;
 	public String[] docnos;
 	public String[] docsContent;
 	
 	public ResultSet(int size){
 		docids = new int[size];
-		scores = new float[size];
+		scores = new double[size];
 		docnos = new String[size];
 		docsContent = new String[size];
 	}
