+++ core/src/main/java/br/ufmg/dcc/latin/reranking/PM2.java
@@ -11,9 +11,9 @@ public class PM2 extends InteractiveReranker {
 	float lambda;
 	int[] highestAspect;
 	
-	private float[] v;
-	private float[] s;
-	private float[][] coverage;
+	private double[] v;
+	private double[] s;
+	private double[][] coverage;
 
 	private String aspectMiningClassName;
 	
@@ -26,9 +26,9 @@ public class PM2 extends InteractiveReranker {
 	
 	public int highestAspect(){
 		int maxQ =  -1;
-		float maxQuotient = -1;
+		double maxQuotient = -1;
 		for (int i = 0; i < v.length; i++) {
-			float quotient = v[i]/(2*s[i]+1);
+			double quotient = v[i]/(2*s[i]+1);
 			if (quotient > maxQuotient) {
 				maxQ = i;
 				maxQuotient = quotient;
@@ -37,22 +37,22 @@ public class PM2 extends InteractiveReranker {
 		return maxQ;
 	}
 	
-	public float score(int docid){
+	public double score(int docid){
 		
 		int q = highestAspect();
 		if (q == -1){
 			return relevance[docid];
 		}
-		float quotientAspectq = v[q]/(2*s[q]+1);
+		double quotientAspectq = v[q]/(2*s[q]+1);
 		quotientAspectq *= coverage[docid][q];
-		float quotientotherAspect  = 0;
+		double quotientotherAspect  = 0;
 		for (int i = 0; i < s.length; i++) {
 			if (i != q) {
 				
 				quotientotherAspect += (v[i]/(2*s[i]+1))*coverage[docid][i];
 			}
 		}
-		float score = lambda*quotientAspectq + (1-lambda)*quotientotherAspect;
+		double score = lambda*quotientAspectq + (1-lambda)*quotientotherAspect;
 		
 		return score;
 	}
@@ -62,12 +62,12 @@ public class PM2 extends InteractiveReranker {
 		if (q == -1) {
 			q = highestAspect();
 		} 
-		float allCoverage = 0;
+		double allCoverage = 0;
 		for (int i = 0; i < coverage[docid].length; ++i) {
 			allCoverage += coverage[docid][i];
 		}
 		if (allCoverage > 0) {
-			float newS = s[q] + coverage[docid][q]/allCoverage;
+			double newS = s[q] + coverage[docid][q]/allCoverage;
 			s[q] = newS;
 		} 
 		highestAspect[docid] = q;
